test(validation): split the citation gate suite and pin the review regressions
Shared fixtures move to citation_freshness_helpers.py and the scope,
diff-parsing, and wiring classes to a sibling file, keeping both under
the file-size ceiling. New cases pin each Copilot finding: the plus-plus
header misread and the U+2028 line shift (both proven by restoring the
defect and watching them fail), line zero, the reasonless marker, the
marker on a removed file, the unrelated indented block, and the finished
neighboring sentence. The wiring test now drives run_all_validations
twice with a monkeypatched validator, failing then passing, so a stub
that stops calling the real gate fails the run rather than the
registration assertion alone.

Refs #5337
